Move post thread metadata from page to layout
